refactor: UserService 인터페이스를 일반 클래스로 통합
구현체가 하나뿐이고 다형성이 필요하지 않아
불필요한 추상화를 제거함
