+++ src/main/java/com/jian/hobbyadventure/service/UserService.java
@@ -1,8 +1,22 @@
 package com.jian.hobbyadventure.service;
 
+import com.jian.hobbyadventure.domain.User;
 import com.jian.hobbyadventure.dto.request.SignupRequest;
+import com.jian.hobbyadventure.repository.UserMapper;
+import lombok.RequiredArgsConstructor;
+import org.springframework.security.crypto.password.PasswordEncoder;
+import org.springframework.stereotype.Service;
 
-public interface UserService {
+@Service
+@RequiredArgsConstructor
+public class UserService {
 
-    void signup(SignupRequest request);
+    private final UserMapper userMapper;
+    private final PasswordEncoder passwordEncoder;
+
+    public void signup(SignupRequest request) {
+        String encodedPassword = passwordEncoder.encode(request.getPassword());
+        User user = new User(request.getEmail(), encodedPassword, request.getNickname());
+        userMapper.insert(user);
+    }
 }
